feat: 补充更多设备类别 + 重构 device_profiles 包结构
1. 新增 3 个设备类别（开关、燃气传感器、门磁传感器）及对应的
   CATEGORY_PROFILES，移除 9 个被类别 profile 完全覆盖的冗余
   MODEL_PROFILES 条目，仅保留 7 个有实质差异的型号 profile。
2. 将 device_profiles.py 单文件（1977 行）重构为 device_profiles/
   包：_constants / _globals / _categories / _models / _resolver，
   对外 API 完全兼容，main.py 零改动。
3. 新增 _discover_devices.py 独立 CLI 工具，支持扫码登录 +
   设备枚举 + miot-spec 属性探测 + profile 模板自动生成。
