+++ README.md
@@ -104,6 +104,9 @@ pip install mijiaAPI
 * `扫地机类别`
 * `热水器类别`
 * `路由器类别`
+* `开关类别`
+* `燃气传感器类别`
+* `门磁传感器类别`
 * `无类别`
 
 说明：
