+++ _conf_schema.json
@@ -17,7 +17,7 @@
     "editor_theme": "vs-dark",
     "obvious_hint": true,
     "default": "{\n  \"客厅 空调\": \"空调类别\",\n  \"卧室 净化器\": \"净化器类别\",\n  \"卧室 温湿度计\": \"温湿度计类别\",\n  \"厨房 热水壶\": \"热水器类别\",\n  \"书房 路由器\": \"路由器类别\"\n}",
-    "hint": "请填写合法 JSON。键(Key)必须与你在 device_map 中配置的设备别名完全一致；值(Value)只能填写以下之一：\"空调类别\"、\"净化器类别\"、\"风扇类别\"、\"蒸煮锅类别\"、\"空气炸锅类别\"、\"温湿度计类别\"、\"体脂秤类别\"、\"扫地机类别\"、\"热水器类别\"、\"路由器类别\"、\"无类别\"。若填写非法值，系统将自动按“无类别”处理。"
+    "hint": "请填写合法 JSON。键(Key)必须与你在 device_map 中配置的设备别名完全一致；值(Value)只能填写以下之一：\"空调类别\"、\"净化器类别\"、\"风扇类别\"、\"蒸煮锅类别\"、\"空气炸锅类别\"、\"温湿度计类别\"、\"体脂秤类别\"、\"扫地机类别\"、\"热水器类别\"、\"路由器类别\"、\"开关类别\"、\"燃气传感器类别\"、\"门磁传感器类别\"、\"无类别\"。"
   },
   "enable_scene_tool": {
     "description": "是否启用米家场景 LLM Tool",
