+++ device_profiles/_categories.py
@@ -1,0 +1,530 @@
+# -*- coding: utf-8 -*-
+"""设备类别模板库。每个类别定义了一组通用的属性/值/展示映射。"""
+
+from ._constants import (
+    CATEGORY_AC,
+    CATEGORY_PURIFIER,
+    CATEGORY_AIR_FRYER,
+    CATEGORY_COOKER,
+    CATEGORY_FAN,
+    CATEGORY_TH_SENSOR,
+    CATEGORY_BODY_SCALE,
+    CATEGORY_VACUUM,
+    CATEGORY_WATER_HEATER,
+    CATEGORY_ROUTER,
+    CATEGORY_SWITCH,
+    CATEGORY_GAS_SENSOR,
+    CATEGORY_DOOR_SENSOR,
+)
+
+CATEGORY_PROFILES = {
+    CATEGORY_AC: {
+        "prop_map": {
+            "温度": "target_temperature",
+            "风速": "fan_level",
+            "模式": "mode",
+            "上下扫风": "vertical_swing",
+            "扫风": "vertical_swing",
+        },
+        "value_map": {
+            "制冷": "cool",
+            "制热": "heat",
+            "除湿": "dry",
+            "送风": "fan",
+            "自动": "auto",
+            "开": True,
+            "关": False,
+            "开扫风": True,
+            "关扫风": False,
+        },
+        "display_map": {
+            "ac_work_mode": "空调工作模式",
+            "vertical_swing": "上下扫风状态",
+            "target_temperature": "设定温度",
+            "ac_state": "空调状态",
+        },
+        "detail_writable": ["on", "mode", "target_temperature", "fan_level", "vertical_swing"],
+        "detail_readable": ["ac_work_mode", "electric_power", "ac_state"],
+        "detail_actions": [],
+        "action_map": {},
+        "help_examples": {
+            "模式": ["制冷", "制热", "自动"],
+            "温度": ["26", "24"],
+            "扫风": ["开", "关"],
+        },
+        "action_examples": [],
+        "help_hints": {
+            "温度": "通常输入 16~30 之间的整数",
+        },
+    },
+
+    CATEGORY_PURIFIER: {
+        "prop_map": {
+            "模式": "mode",
+            "风速": "fan_level",
+            "童锁": "physical_controls_locked",
+            "提示音": "alarm",
+            "屏幕": "brightness",
+        },
+        "value_map": {
+            "自动": 0,
+            "睡眠": 1,
+            "最爱": 2,
+            "亮": 0,
+            "暗": 1,
+            "熄灭": 2,
+        },
+        "detail_writable": ["on", "mode", "fan_level", "physical_controls_locked", "alarm", "brightness"],
+        "detail_readable": ["temperature", "relative_humidity", "pm2.5_density", "filter_left_time", "filter_life_level"],
+        "detail_actions": [],
+        "action_map": {},
+        "help_examples": {
+            "模式": ["自动", "睡眠", "最爱"],
+            "童锁": ["开", "关"],
+            "屏幕": ["亮", "暗", "熄灭"],
+            "提示音": ["开", "关"],
+        },
+        "action_examples": [],
+        "help_hints": {},
+    },
+
+    CATEGORY_AIR_FRYER: {
+        "prop_map": {
+            "模式": "mode",
+            "时间": "target_time",
+            "温度": "target_temperature",
+            "自动保温": "auto_keep_warm",
+            "翻面提醒": "turn_pot_config",
+            "口感": "texture",
+            "重量": "cooking_weight",
+        },
+        "value_map": {
+            "开": True,
+            "关": False,
+        },
+        "display_map": {
+            "target_time": "设定时间",
+            "target_temperature": "设定温度",
+            "recipe_name": "当前食谱",
+            "turn_pot": "翻锅提醒状态",
+            "current_keep_warm": "当前正处保温",
+            "reservation_left_time": "距离预约开始",
+            "texture": "口感设置",
+        },
+        "detail_writable": [
+            "on", "mode", "target_time", "target_temperature",
+            "auto_keep_warm", "turn_pot_config", "texture", "cooking_weight",
+        ],
+        "detail_readable": ["status", "left_time", "recipe_name", "current_keep_warm", "reservation_left_time"],
+        "detail_actions": [],
+        "action_map": {},
+        "help_examples": {
+            "温度": ["180", "200"],
+            "时间": ["15", "20"],
+            "翻面提醒": ["开", "关"],
+        },
+        "action_examples": [],
+        "help_hints": {
+            "时间": "输入预计分钟数",
+        },
+    },
+
+    CATEGORY_COOKER: {
+        "prop_map": {
+            "模式": "mode",
+            "时间": "target_time",
+            "温度": "target_temperature",
+            "火力": "heat_level",
+            "保温": "keepwarm_set",
+            "提示音": "alarm",
+            "预约": "cookreservation",
+            "暂停": "switch_pausetoadd",
+        },
+        "value_map": {
+            "开": True,
+            "关": False,
+        },
+        "display_map": {
+            "target_time": "设定时间",
+            "target_temperature": "设定温度",
+            "temperature": "锅内实时温度",
+            "cook_done": "烹饪是否完成",
+            "recipename": "当前食谱",
+            "warm_temperature": "设定保温温度",
+            "switch_pausetoadd": "中途加料开关",
+        },
+        "detail_writable": [
+            "on", "mode", "target_time", "target_temperature",
+            "heat_level", "keepwarm_set", "alarm",
+            "cookreservation", "switch_pausetoadd",
+        ],
+        "detail_readable": ["status", "left_time", "temperature", "cook_done", "recipename"],
+        "detail_actions": [],
+        "action_map": {},
+        "help_examples": {
+            "火力": ["1", "3", "5"],
+            "暂停": ["开", "关"],
+            "保温": ["开", "关"],
+        },
+        "action_examples": [],
+        "help_hints": {},
+    },
+
+    CATEGORY_FAN: {
+        "prop_map": {
+            "模式": "mode",
+            "风速": "fan_level",
+            "摇头": "horizontal_swing",
+            "童锁": "physical_controls_locked",
+        },
+        "value_map": {
+            "直吹": "straight",
+            "自然风": "nature",
+            "开": True,
+            "关": False,
+        },
+        "detail_writable": ["on", "mode", "fan_level", "horizontal_swing", "physical_controls_locked"],
+        "detail_readable": [],
+        "detail_actions": [],
+        "action_map": {},
+        "help_examples": {
+            "模式": ["直吹", "自然风"],
+            "风速": ["1", "2", "3", "4"],
+            "摇头": ["开", "关"],
+        },
+        "action_examples": [],
+        "help_hints": {},
+    },
+
+    CATEGORY_TH_SENSOR: {
+        "prop_map": {},
+        "value_map": {},
+        "display_map": {
+            "battery_level": "电池电量",
+            "temperature": "当前温度",
+            "relative_humidity": "当前湿度",
+        },
+        "detail_writable": [],
+        "detail_readable": ["temperature", "relative_humidity", "battery_level"],
+        "detail_actions": [],
+        "action_map": {},
+        "help_examples": {},
+        "action_examples": [],
+        "help_hints": {},
+    },
+
+    CATEGORY_BODY_SCALE: {
+        "prop_map": {},
+        "value_map": {},
+        "display_map": {
+            "fault": "故障状态",
+            "battery": "电池状态",
+            "unit": "单位",
+            "mode": "工作模式",
+            "status": "工作状态",
+        },
+        "detail_writable": [],
+        "detail_readable": ["fault", "battery", "unit", "mode", "status"],
+        "detail_actions": [],
+        "action_map": {},
+        "help_examples": {},
+        "action_examples": [],
+        "help_hints": {},
+    },
+
+    CATEGORY_VACUUM: {
+        "prop_map": {
+            "吸力": "mode",
+            "模式": "mode",
+            "扫拖模式": "sweep_mop_type",
+            "清扫方式": "sweep_type",
+            "拖地水量": "mop_water_output_level",
+            "提示音": "alarm",
+            "音量": "volume",
+            "童锁": "physical_controls_locked",
+            "地毯增压": "carpet_boost",
+            "地毯避让": "carpet_avoidance",
+            "断点续扫": "sweep_break_switch",
+            "回洗拖布": "enable_mop_wash",
+            "回洗频率": "frequency_mop_wash",
+            "自动集尘": "auto_dust_arrest",
+            "集尘频率": "dust_arrest_frequency",
+            "自动烘干拖布": "auto_mop_dry",
+            "自动上下水": "auto_water_change",
+            "使用清洁剂": "use_detergent",
+            "AI清洁": "ai_cleaning",
+            "热水洗拖布": "hot_water_mop_wash",
+            "拖布自动抬升": "mop_auto_lift",
+            "勿扰": "no_disturb",
+        },
+        "value_map": {
+            "开": True,
+            "关": False,
+            "开启": True,
+            "关闭": False,
+            "安静": 1,
+            "标准": 2,
+            "强力": 3,
+            "超强": 4,
+            "仅扫地": 1,
+            "仅拖地": 2,
+            "扫拖": 3,
+            "先扫后拖": 4,
+            "全屋清洁": 1,
+            "建图": 2,
+            "选区清扫": 3,
+            "沿边模式": 4,
+            "定制清扫": 5,
+            "指哪扫哪": 6,
+            "关闭": 0,
+            "1挡": 1,
+            "2挡": 2,
+            "3挡": 3,
+            "按房间": 0,
+            "按面积": 1,
+            "按时间": 2,
+            "2小时": 1,
+            "3小时": 2,
+            "4小时": 3,
+        },
+        "display_map": {
+            "status": "工作状态",
+            "fault": "故障状态",
+            "battery_level": "电池电量",
+            "charging_state": "充电状态",
+            "cleaning_area": "当前清扫面积",
+            "cleaning_time": "当前清扫时长",
+            "clean_times": "清洁次数",
+            "mode": "吸力模式",
+            "sweep_mop_type": "扫拖模式",
+            "sweep_type": "清扫方式",
+            "mop_water_output_level": "拖地水量",
+            "mop_status": "拖布状态",
+            "carpet_boost": "地毯增压模式",
+            "carpet_avoidance": "地毯避让",
+            "enable_mop_wash": "回洗拖布",
+            "frequency_mop_wash": "回洗频率",
+            "auto_dust_arrest": "自动集尘",
+            "auto_mop_dry": "自动烘干拖布",
+            "no_disturb": "勿扰模式",
+            "mop_life_level": "拖布剩余寿命",
+            "mop_left_time": "拖布剩余时间",
+            "brush_life_level": "主刷剩余寿命",
+            "brush_left_time": "主刷剩余时间",
+            "filter_life_level": "滤芯剩余寿命",
+            "filter_left_time": "滤芯剩余时间",
+            "detergent_left_level": "清洁剂剩余量",
+            "dust_bag_life_level": "尘袋剩余寿命",
+            "dust_bag_left_time": "尘袋剩余时间",
+            "cleaning_progress": "清扫进度",
+            "drying_progress": "烘干进度",
+            "dry_left_time": "烘干剩余时长",
+        },
+        "detail_writable": [
+            "mode", "sweep_mop_type", "sweep_type",
+            "mop_water_output_level", "alarm", "volume",
+            "physical_controls_locked", "carpet_boost", "carpet_avoidance",
+            "sweep_break_switch", "enable_mop_wash", "frequency_mop_wash",
+            "auto_dust_arrest", "dust_arrest_frequency",
+            "auto_mop_dry", "auto_water_change", "use_detergent",
+            "ai_cleaning", "hot_water_mop_wash", "mop_auto_lift", "no_disturb",
+        ],
+        "detail_readable": [
+            "status", "fault", "battery_level", "charging_state",
+            "cleaning_area", "cleaning_time", "clean_times",
+            "cleaning_progress", "drying_progress", "dry_left_time",
+            "mop_status", "mop_life_level", "mop_left_time",
+            "brush_life_level", "brush_left_time",
+            "filter_life_level", "filter_left_time",
+            "detergent_left_level",
+            "dust_bag_life_level", "dust_bag_left_time",
+            "current_no_disturb",
+        ],
+        "action_map": {
+            "开始清扫": "start_sweep",
+            "暂停": "pause_sweeping",
+            "继续": "continue_sweep",
+            "回充": "start_charge",
+            "开始拖地": "start_mop",
+            "开始扫拖": "start_sweep_mop",
+            "开始建图": "start_build_map",
+            "开始集尘": "start_dust_arrest",
+            "开始洗拖布": "start_mop_wash",
+            "开始烘干": "start_dry",
+            "停止烘干": "stop_dry",
+            "停止洗拖布": "stop_mop_wash",
+            "停止并回充": "stop_and_gocharge",
+        },
+        "detail_actions": [
+            "start_sweep", "pause_sweeping", "continue_sweep",
+            "start_charge", "start_mop", "start_sweep_mop",
+            "start_build_map", "start_dust_arrest", "start_mop_wash",
+            "start_dry", "stop_dry", "stop_mop_wash", "stop_and_gocharge",
+        ],
+        "help_examples": {
+            "吸力": ["安静", "标准", "强力", "超强"],
+            "扫拖模式": ["仅扫地", "仅拖地", "扫拖", "先扫后拖"],
+            "清扫方式": ["全屋清洁", "选区清扫", "沿边模式", "建图"],
+            "拖地水量": ["关闭", "1挡", "2挡", "3挡"],
+            "地毯增压": ["开", "关"],
+            "地毯避让": ["开", "关"],
+            "童锁": ["开", "关"],
+            "提示音": ["开", "关"],
+            "勿扰": ["开", "关"],
+        },
+        "action_examples": [
+            "开始清扫", "暂停", "继续", "回充", "开始拖地", "开始扫拖",
+        ],
+        "help_hints": {
+            "音量": "支持 0~100",
+        },
+    },
+
+    CATEGORY_WATER_HEATER: {
+        "prop_map": {
+            "温度": "target_temperature",
+            "加热温度": "target_temperature",
+            "保温温度": "keep_warm_temperature",
+            "定温时长": "keep_warm_time",
+            "自动保温": "auto_keep_warm",
+            "勿扰": "no_disturb",
+            "提壶记忆": "lift_remember_temp",
+            "烧开提醒": "boiling_reminder",
+            "定温提醒": "keep_warm_reminder",
+        },
+        "value_map": {
+            "开": True,
+            "关": False,
+            "开启": True,
+            "关闭": False,
+        },
+        "display_map": {
+            "status": "工作状态",
+            "fault": "故障状态",
+            "temperature": "当前水温",
+            "target_temperature": "加热温度",
+            "keep_warm_temperature": "保温温度",
+            "auto_keep_warm": "自动保温",
+            "on": "开关状态",
+            "no_disturb": "勿扰模式",
+            "keep_warm_time": "定温时长",
+            "lift_remember_temp": "提壶记忆定温",
+            "boiling_reminder": "烧开提醒",
+            "keep_warm_reminder": "定温提醒",
+            "kettle_lifting": "提壶状态",
+            "warming_time": "已定温时长",
+            "target_mode": "记忆工作模式",
+            "heat_mode": "加热模式",
+        },
+        "detail_writable": [
+            "on", "target_temperature", "auto_keep_warm",
+            "keep_warm_temperature", "no_disturb", "keep_warm_time",
+            "lift_remember_temp", "boiling_reminder", "keep_warm_reminder",
+            "target_mode", "heat_mode",
+        ],
+        "detail_readable": [
+            "status", "fault", "temperature", "target_temperature",
+            "auto_keep_warm", "keep_warm_temperature",
+            "on", "no_disturb", "keep_warm_time",
+            "kettle_lifting", "warming_time",
+        ],
+        "detail_actions": ["stop_work", "push_water_reminder"],
+        "action_map": {
+            "停止工作": "stop_work",
+            "喝水提醒": "push_water_reminder",
+            "推送喝水提醒": "push_water_reminder",
+        },
+        "help_examples": {
+            "温度": ["70", "85", "95"],
+            "保温温度": ["45", "55", "70"],
+            "定温时长": ["60", "120", "1440"],
+            "自动保温": ["开", "关"],
+            "提壶记忆": ["开", "关"],
+            "烧开提醒": ["开", "关"],
+            "定温提醒": ["开", "关"],
+            "勿扰": ["开", "关"],
+        },
+        "action_examples": ["停止工作", "喝水提醒"],
+        "help_hints": {
+            "温度": "支持 40~99°C",
+            "保温温度": "支持 40~90°C",
+            "定温时长": "支持 60~1440 分钟",
+        },
+    },
+
+    CATEGORY_ROUTER: {
+        "prop_map": {},
+        "value_map": {},
+        "display_map": {
+            "download_speed": "下载速率",
+            "connected_device_number": "连接设备数",
+        },
+        "detail_writable": [],
+        "detail_readable": ["download_speed", "connected_device_number"],
+        "detail_actions": [],
+        "action_map": {},
+        "help_examples": {},
+        "action_examples": [],
+        "help_hints": {},
+    },
+
+    CATEGORY_SWITCH: {
+        "prop_map": {
+            "开关": "on",
+            "模式": "mode",
+        },
+        "value_map": {},
+        "display_map": {
+            "on": "电源状态",
+            "mode": "运行模式",
+            "anti_flicker": "防闪烁",
+            "temperature": "模块温度",
+        },
+        "detail_writable": ["on", "mode", "anti_flicker"],
+        "detail_readable": ["on", "mode", "anti_flicker", "temperature"],
+        "action_map": {
+            "切换": "toggle",
+        },
+        "detail_actions": ["toggle"],
+        "help_examples": {
+            "开关": ["开", "关"],
+        },
+        "action_examples": ["切换"],
+        "help_hints": {},
+    },
+
+    CATEGORY_DOOR_SENSOR: {
+        "prop_map": {},
+        "value_map": {},
+        "display_map": {
+            "contact_state": "门窗状态",
+            "illumination": "光照强度",
+            "battery_level": "电池电量",
+        },
+        "detail_writable": [],
+        "detail_readable": ["contact_state", "illumination", "battery_level"],
+        "detail_actions": [],
+        "action_map": {},
+        "help_examples": {},
+        "action_examples": [],
+        "help_hints": {},
+    },
+
+    CATEGORY_GAS_SENSOR: {
+        "prop_map": {
+            "消音": "mute",
+        },
+        "value_map": {},
+        "display_map": {
+            "status": "报警状态",
+            "gas_concentration": "天然气浓度(%LEL)",
+        },
+        "detail_writable": ["self_inspection", "mute"],
+        "detail_readable": ["status", "gas_concentration"],
+        "detail_actions": [],
+        "action_map": {},
+        "help_examples": {},
+        "action_examples": [],
+        "help_hints": {},
+    },
+}
