+++ device_profiles/_constants.py
@@ -1,0 +1,33 @@
+# -*- coding: utf-8 -*-
+
+CATEGORY_NONE = "无类别"
+CATEGORY_AC = "空调类别"
+CATEGORY_PURIFIER = "净化器类别"
+CATEGORY_FAN = "风扇类别"
+CATEGORY_COOKER = "蒸煮锅类别"
+CATEGORY_AIR_FRYER = "空气炸锅类别"
+CATEGORY_TH_SENSOR = "温湿度计类别"
+CATEGORY_BODY_SCALE = "体脂秤类别"
+CATEGORY_VACUUM = "扫地机类别"
+CATEGORY_WATER_HEATER = "热水器类别"
+CATEGORY_ROUTER = "路由器类别"
+CATEGORY_SWITCH = "开关类别"
+CATEGORY_GAS_SENSOR = "燃气传感器类别"
+CATEGORY_DOOR_SENSOR = "门磁传感器类别"
+
+VALID_CATEGORIES = {
+    CATEGORY_NONE,
+    CATEGORY_AC,
+    CATEGORY_PURIFIER,
+    CATEGORY_FAN,
+    CATEGORY_COOKER,
+    CATEGORY_AIR_FRYER,
+    CATEGORY_TH_SENSOR,
+    CATEGORY_BODY_SCALE,
+    CATEGORY_VACUUM,
+    CATEGORY_WATER_HEATER,
+    CATEGORY_ROUTER,
+    CATEGORY_SWITCH,
+    CATEGORY_GAS_SENSOR,
+    CATEGORY_DOOR_SENSOR,
+}
