+++ device_profiles/_globals.py
@@ -1,0 +1,141 @@
+# -*- coding: utf-8 -*-
+"""全局兜底映射表，作为类别/型号 profile 未覆盖时的最终 fallback。"""
+
+# 控制用: 中文 -> 英文属性名
+GLOBAL_PROP_MAP = {
+    "开关": "on",
+    "温度": "target_temperature",
+    "风速": "fan_level",
+    "模式": "mode",
+    "亮度": "brightness",
+}
+
+# 控制用: 中文值 -> 实际下发值
+GLOBAL_VAL_MAP = {
+    "低": "low",
+    "中": "medium",
+    "高": "high",
+    "一档": 1,
+    "二档": 2,
+    "三档": 3,
+}
+
+# 展示用: 英文属性名 -> 中文友好名
+GLOBAL_DISPLAY_MAP = {
+    "temperature": "当前温度",
+    "relative_humidity": "当前湿度",
+    "pm2.5_density": "PM2.5浓度",
+    "filter_left_time": "滤芯剩余天数",
+    "filter_life_level": "滤芯寿命百分比",
+    "filter_used_time": "滤芯已使用时长",
+    "mode": "运行模式",
+    "fan_level": "风速档位",
+    "on": "电源状态",
+    "status": "当前状态",
+    "left_time": "剩余时间",
+    "physical_controls_locked": "童锁状态",
+    "brightness": "屏幕亮度",
+    "alarm": "提示音状态",
+    "target_time": "设定的目标时间",
+    "heat_level": "当前火力",
+    "keepwarm_set": "保温设置",
+    "electric_power": "功率",
+    "target_temperature": "设定温度",
+    "ac_state": "空调状态",
+    "ac_work_mode": "空调工作模式",
+    "vertical_swing": "上下扫风状态",
+    "horizontal_swing": "摇头状态",
+    "recipe_name": "当前食谱",
+    "recipeid": "食谱编号",
+    "recipe_id": "食谱编号",
+    "recipename": "当前食谱",
+    "current_keep_warm": "当前保温状态",
+    "reservation_left_time": "距离预约开始",
+    "texture": "口感设置",
+    "warm_temperature": "保温温度",
+    "cook_done": "烹饪是否完成",
+    "switch_pausetoadd": "中途加料",
+    "cookreservation": "预约设置",
+    "air_quality": "空气质量状态",
+    "fault": "故障状态",
+    "moto_speed_rpm": "电机转速",
+    "battery_level": "电池电量",
+    "battery": "电池状态",
+    "turn_pot": "翻面提醒状态",
+    "auto_keep_warm": "自动保温",
+    "power_set": "功率设置",
+    "temp_reached": "到温提醒",
+    "temp_reached_set": "提醒温度",
+    "target_cooking_measure": "食材份量",
+    "cooking_weight": "食材重量",
+    "match_state": "匹配状态",
+    "unit": "单位",
+
+    # 热水器 / 电水壶
+    "keep_warm_temperature": "保温温度",
+    "keep_warm_time": "定温时长",
+    "warming_time": "已定温时长",
+    "lift_remember_temp": "提壶记忆定温",
+    "boiling_reminder": "烧开提醒",
+    "keep_warm_reminder": "定温提醒",
+    "kettle_lifting": "提壶状态",
+    "target_mode": "记忆工作模式",
+    "heat_mode": "加热模式",
+    "no_disturb": "勿扰模式",
+
+    # 路由器
+    "download_speed": "下载速率",
+    "connected_device_number": "连接设备数",
+
+    # 扫地机
+    "charging_state": "充电状态",
+    "cleaning_area": "当前清扫面积",
+    "cleaning_time": "当前清扫时长",
+    "clean_times": "清洁次数",
+    "sweep_mop_type": "扫拖模式",
+    "sweep_type": "清扫方式",
+    "mop_water_output_level": "拖地水量",
+    "mop_status": "拖布状态",
+    "carpet_boost": "地毯增压模式",
+    "carpet_avoidance": "地毯避让",
+    "carpet_display": "地毯显示",
+    "sweep_break_switch": "断点续扫",
+    "sleep_status": "休眠状态",
+    "location_status": "定位状态",
+    "enable_mop_wash": "回洗拖布",
+    "frequency_mop_wash": "回洗频率",
+    "water_output_for_washing_mop": "拖布出水量",
+    "drying_time": "烘干时长",
+    "auto_dust_arrest": "自动集尘",
+    "dust_arrest_frequency": "集尘频率",
+    "auto_mop_dry": "自动烘干拖布",
+    "auto_water_change": "自动上下水",
+    "use_detergent": "使用清洁剂",
+    "ai_cleaning": "AI感知清洁",
+    "hot_water_mop_wash": "热水洗拖布",
+    "mop_auto_lift": "拖布自动抬升",
+    "dirt_detection": "脏污识别",
+    "object_detection": "物体识别",
+    "room_detection": "房间识别",
+    "floor_material_detection": "地板材质识别",
+    "water_check_status": "上下水检测状态",
+    "detergent_self_delivery": "清洁剂自投放",
+    "current_physical_control_lock": "物理控制锁状态",
+    "mop_life_level": "拖布剩余寿命",
+    "mop_left_time": "拖布剩余时间",
+    "brush_life_level": "主刷剩余寿命",
+    "brush_left_time": "主刷剩余时间",
+    "current_no_disturb": "当前勿扰状态",
+    "detergent_left_level": "清洁剂剩余量",
+    "detergent_left_time": "清洁剂剩余时间",
+    "dust_bag_life_level": "尘袋剩余寿命",
+    "dust_bag_left_time": "尘袋剩余时间",
+    "download_status": "下载状态",
+    "download_progress": "下载进度",
+    "drying_progress": "烘干进度",
+    "cleaning_progress": "清扫进度",
+    "dry_left_time": "烘干剩余时长",
+    "statistical_clean_area": "累计清洁面积",
+    "statistical_cleaning_times": "累计清洁次数",
+    "statistical_cleaning_duration": "累计清洁时长",
+}
