+++ device_profiles/_models.py
@@ -1,0 +1,697 @@
+# -*- coding: utf-8 -*-
+"""
+型号精确模板库。
+
+仅收录与对应类别 profile 有显著差异的设备型号。
+如果某型号的属性/值/动作完全被类别 profile 覆盖，则无需在此注册。
+"""
+
+from typing import Any, Dict
+
+from ._constants import (
+    CATEGORY_AC,
+    CATEGORY_AIR_FRYER,
+    CATEGORY_BODY_SCALE,
+    CATEGORY_COOKER,
+    CATEGORY_PURIFIER,
+    CATEGORY_VACUUM,
+    CATEGORY_WATER_HEATER,
+)
+
+MODEL_PROFILES: Dict[str, Dict[str, Any]] = {
+
+    # ---- 空气炸锅 ----
+    "careli.fryer.maf07c": {
+        "category": CATEGORY_AIR_FRYER,
+        "hidden_props": [],
+        "prop_map": {
+            "模式": "mode",
+            "时间": "target_time",
+            "温度": "target_temperature",
+            "自动保温": "auto_keep_warm",
+            "当前保温": "current_keep_warm",
+            "翻面提醒": "turn_pot_config",
+            "口感": "texture",
+            "重量": "cooking_weight",
+            "预约": "reservation_left_time",
+            "食谱名": "recipe_name",
+            "食谱": "recipe_name",
+        },
+        "value_map": {
+            "开": True,
+            "关": False,
+            "薯条": 1,
+            "鸡翅": 2,
+            "牛排": 3,
+            "羊排": 4,
+            "五花肉": 5,
+            "鱼": 6,
+            "红薯": 7,
+            "蛋糕": 8,
+            "蛋挞": 9,
+            "解冻": 10,
+            "果干": 11,
+            "酸奶": 12,
+            "手动模式": 0,
+            "手动": 0,
+            "脆烤": 1,
+            "嫩烤": 2,
+            "脱脂": 3,
+        },
+        "display_map": {
+            "status": "工作状态",
+            "fault": "故障状态",
+            "target_time": "设定时间",
+            "target_temperature": "设定温度",
+            "left_time": "剩余时间",
+            "auto_keep_warm": "自动保温",
+            "current_keep_warm": "当前保温状态",
+            "mode": "烹饪模式",
+            "recipe_id": "食谱编号",
+            "recipe_name": "食谱名称",
+            "target_cooking_measure": "食材份量",
+            "turn_pot": "翻面提醒状态",
+            "turn_pot_config": "翻面提醒设置",
+            "texture": "口感设置",
+            "reservation_left_time": "预约剩余时间",
+            "cooking_weight": "食材重量",
+        },
+        "detail_writable": [
+            "mode", "target_time", "target_temperature",
+            "auto_keep_warm", "turn_pot_config", "texture",
+            "reservation_left_time", "cooking_weight",
+        ],
+        "detail_readable": [
+            "status", "fault", "left_time",
+            "current_keep_warm", "turn_pot",
+            "reservation_left_time", "recipe_name",
+        ],
+        "action_map": {},
+        "detail_actions": [],
+        "help_examples": {
+            "模式": ["薯条", "鸡翅", "牛排", "手动"],
+            "温度": ["180", "200"],
+            "时间": ["15", "20"],
+            "翻面提醒": ["开", "关"],
+            "口感": ["脆烤", "嫩烤", "脱脂"],
+        },
+        "action_examples": [],
+        "help_hints": {
+            "温度": "支持 40~220，步进 5",
+            "时间": "支持 1~1440 分钟",
+            "重量": "支持 100~1800 克，步进 100",
+        },
+    },
+
+    # ---- 蒸煮锅 ----
+    "deerma.mfcp.v3": {
+        "category": CATEGORY_COOKER,
+        "hidden_props": [
+            "recipe_command", "recipe_sort", "mode_sort",
+            "recipes_recipe_command", "stepid",
+            "user_device_info", "lifetime",
+        ],
+        "prop_map": {
+            "模式": "mode",
+            "时间": "target_time",
+            "温度": "target_temperature",
+            "火力": "heat_level",
+            "提示音": "alarm",
+            "功率": "power_set",
+            "到温提醒": "temp_reached",
+            "提醒温度": "temp_reached_set",
+            "预约": "cookreservation",
+            "保温": "keepwarm_set",
+            "保温温度": "warm_temperature",
+            "加料提醒": "switch_pausetoadd",
+        },
+        "value_map": {
+            "开": True,
+            "关": False,
+            "火锅": 1,
+            "蒸煮": 2,
+            "煲汤": 3,
+            "煮粥": 4,
+            "煮面": 5,
+            "保温": 6,
+            "安全模式": 7,
+            "食谱": 8,
+            "预约模式": 9,
+        },
+        "display_map": {
+            "status": "工作状态",
+            "mode": "工作模式",
+            "left_time": "剩余时间",
+            "fault": "故障状态",
+            "heat_level": "加热档位",
+            "target_temperature": "设定温度",
+            "temperature": "当前温度",
+            "target_time": "设定时间",
+            "alarm": "提示音状态",
+            "power_set": "功率设置",
+            "temp_reached": "到温提醒",
+            "cook_done": "完成提醒",
+            "temp_reached_set": "提醒温度",
+            "recipeid": "菜谱编号",
+            "cookreservation": "预约烹饪时间",
+            "keepwarm_set": "烹饪后保温",
+            "warm_temperature": "保温设定温度",
+            "recipename": "菜谱名称",
+            "switch_pausetoadd": "加料提醒",
+        },
+        "detail_writable": [
+            "mode", "target_time", "target_temperature",
+            "heat_level", "alarm", "power_set",
+            "temp_reached", "cook_done", "temp_reached_set",
+            "cookreservation", "keepwarm_set",
+            "warm_temperature", "switch_pausetoadd",
+        ],
+        "detail_readable": [
+            "status", "left_time", "fault", "temperature", "recipename",
+        ],
+        "action_map": {},
+        "detail_actions": [],
+        "help_examples": {
+            "模式": ["火锅", "蒸煮", "煲汤", "煮粥", "煮面", "保温"],
+            "火力": ["1", "3", "5", "9"],
+            "温度": ["60", "80", "100"],
+            "时间": ["10", "30", "60"],
+            "保温": ["开", "关"],
+        },
+        "action_examples": [],
+        "help_hints": {
+            "火力": "支持 1~9 档",
+            "温度": "支持 30~100°C",
+            "时间": "支持 0~300 分钟",
+        },
+    },
+
+    # ---- 空调伴侣 ----
+    "lumi.acpartner.mcn02": {
+        "category": CATEGORY_AC,
+        "hidden_props": [
+            "ac_ctrl_library", "ac_ctrl_library_crc32",
+            "ac_type", "brand_id", "remote_id",
+        ],
+        "prop_map": {
+            "开关": "on",
+            "模式": "mode",
+            "温度": "target_temperature",
+            "风速": "fan_level",
+            "上下扫风": "vertical_swing",
+            "扫风": "vertical_swing",
+        },
+        "value_map": {
+            "开": True,
+            "关": False,
+            "自动": 0,
+            "制冷": 1,
+            "除湿": 2,
+            "制热": 3,
+            "送风": 4,
+            "低": 1,
+            "中": 2,
+            "高": 3,
+        },
+        "display_map": {
+            "on": "开关状态",
+            "mode": "模式",
+            "target_temperature": "设定温度",
+            "fan_level": "风速档位",
+            "vertical_swing": "上下摆风",
+            "match_state": "匹配状态",
+            "ac_work_mode": "空调工作模式",
+            "ac_state": "空调状态",
+            "electric_power": "电功率",
+        },
+        "detail_writable": [
+            "on", "mode", "target_temperature", "fan_level", "vertical_swing",
+        ],
+        "detail_readable": [
+            "ac_work_mode", "ac_state", "electric_power", "match_state",
+        ],
+        "action_map": {},
+        "detail_actions": [],
+        "help_examples": {
+            "模式": ["自动", "制冷", "除湿", "制热", "送风"],
+            "温度": ["26", "24"],
+            "风速": ["低", "中", "高"],
+            "扫风": ["开", "关"],
+        },
+        "action_examples": [],
+        "help_hints": {
+            "温度": "支持 16~30°C",
+        },
+    },
+
+    # ---- 体脂秤 (保留：大量 hidden_props) ----
+    "yunmai.scales.ms103": {
+        "category": CATEGORY_BODY_SCALE,
+        "hidden_props": [
+            "app_version", "balance_time", "body_composition",
+            "current_time", "customized_argument_3",
+            "customized_argument_4", "customized_argument_5",
+            "firmwareversion", "hash_code", "heart_rate_detection",
+            "idx", "idx_string", "is_owner", "mac", "miid",
+            "offline_data_count", "offline_msg_cnt",
+            "only_upload_count", "pluginversion", "real_weight",
+            "report_offline_data", "self_check_items",
+            "self_check_result", "share_user_list", "sn",
+            "syncinfo_current_time", "syncinfo_heart_rate_detection",
+            "syncinfo_unit", "time", "user_age", "user_bfp",
+            "user_count", "user_height", "user_id", "user_sex",
+            "user_status", "user_type", "user_weight",
+            "userinfolist", "weight_data",
+        ],
+        "prop_map": {},
+        "value_map": {},
+        "display_map": {
+            "fault": "故障状态",
+            "battery": "电池状态",
+            "unit": "单位",
+            "mode": "工作模式",
+            "status": "工作状态",
+        },
+        "detail_writable": [],
+        "detail_readable": ["fault", "battery", "unit", "mode", "status"],
+        "action_map": {},
+        "detail_actions": [],
+        "help_examples": {},
+        "action_examples": [],
+        "help_hints": {},
+    },
+
+    # ---- 空气净化器 (保留：特有亮度档位 + 风速范围) ----
+    "zhimi.airp.rma3": {
+        "category": CATEGORY_PURIFIER,
+        "hidden_props": [
+            "aqi_updata_heartbeat", "filter_used_time_dbg", "moto_speed_rpm",
+        ],
+        "prop_map": {
+            "开关": "on",
+            "模式": "mode",
+            "风速": "fan_level",
+            "童锁": "physical_controls_locked",
+            "提示音": "alarm",
+            "屏幕": "brightness",
+            "亮度": "brightness",
+        },
+        "value_map": {
+            "自动": 0,
+            "睡眠": 1,
+            "最爱": 2,
+            "开": True,
+            "关": False,
+            "开启": True,
+            "关闭": False,
+            "息屏": 0,
+            "微亮": 1,
+            "正常": 2,
+            "熄灭": 0,
+            "暗": 1,
+            "亮": 2,
+        },
+        "display_map": {
+            "on": "开关状态",
+            "mode": "运行模式",
+            "fan_level": "风速",
+            "physical_controls_locked": "童锁状态",
+            "alarm": "提示音状态",
+            "brightness": "屏幕亮度",
+            "temperature": "当前温度",
+            "relative_humidity": "当前湿度",
+            "pm2.5_density": "PM2.5浓度",
+            "air_quality": "空气质量状态",
+            "fault": "故障状态",
+            "filter_left_time": "滤芯剩余天数",
+            "filter_life_level": "滤芯寿命百分比",
+            "filter_used_time": "滤芯已使用时长",
+        },
+        "detail_writable": [
+            "on", "mode", "fan_level",
+            "physical_controls_locked", "alarm", "brightness",
+        ],
+        "detail_readable": [
+            "pm2.5_density", "temperature", "relative_humidity",
+            "air_quality", "filter_left_time",
+            "filter_life_level", "filter_used_time", "fault",
+        ],
+        "action_map": {},
+        "detail_actions": [],
+        "help_examples": {
+            "模式": ["自动", "睡眠", "最爱"],
+            "童锁": ["开", "关"],
+            "提示音": ["开", "关"],
+            "屏幕": ["息屏", "微亮", "正常"],
+            "风速": ["0", "1", "5", "10", "14"],
+        },
+        "action_examples": [],
+        "help_hints": {
+            "风速": '可输入 0~14 的整数档位；通常仅在"最爱"模式下更有意义',
+            "屏幕": "支持：息屏 / 微亮 / 正常",
+            "模式": "支持：自动 / 睡眠 / 最爱",
+        },
+    },
+
+    # ---- 扫地机 (保留：大量 hidden_props + 扩展值映射) ----
+    "xiaomi.vacuum.ov21cn": {
+        "category": CATEGORY_VACUUM,
+        "hidden_props": [
+            "vacuum_frameware_version", "zone_ids",
+            "restricted_sweep_areas", "restricted_walls",
+            "vacuum_room_ids", "room_information", "last_clean_time",
+            "base_station_working_status", "order_clean",
+            "common_params", "map_complete_dialog",
+            "exit_remote_dialog", "button_type",
+            "current_cleaning_config", "user_define_sweep_cfg",
+            "user_define_sweep_id", "sweep_ai_detection",
+            "liquid_dirt_detection", "solid_dirt_detection",
+            "water_check_list", "edge_swing_tail_sweep",
+            "sweep_ai_object", "sweep_furniture", "carpet_object",
+            "mop_water_output_level_no_tank",
+            "frequency_mop_wash_no_tank",
+            "auto_water_change_installed", "cut_hair_config",
+            "frequency_cut_hair_config", "vacuum_route",
+            "fault_ids", "action_result", "edge_sweep_frequency",
+            "ai_managed_cleaning", "plugin_info_remind",
+            "detergent_depletion_reminder", "notice",
+            "carpet_cleaning_method", "sweep_route",
+            "obstacle_avoidance_strategy", "carpet_deep_cleaning",
+            "carpet_discriminate", "sweep_carpet_first",
+            "mop_outer_swing", "mop_outer_swing_frequency",
+            "frequency_mop_wash_by_time",
+            "auto_dust_arrest_power_level",
+            "worry_free_clean_mode", "wash_mop_water_temperature",
+            "brush_outswing", "repeat_wash_mop",
+            "real_scene_capture", "auto_night_fill_light",
+            "ai_wake_up", "sweep_mop_status",
+            "sewage_tank_status", "water_tank_status", "sill",
+            "base_station_water_tank_status",
+            "host_water_tank_status", "furniture_for_automation",
+            "map_hidden_zone", "voltage", "map_obj_name",
+            "trajectory_obj_name", "clean_record",
+            "vacuum_position", "map_management", "current_map_id",
+            "carpet_obj_name", "vacuum_map_common_params",
+            "map_3d_info", "map_switch_method", "map_upload_stop",
+            "map_upload_stop_action_result", "backup_map_list",
+            "restore_map_result", "cloud_clean_record",
+            "enable_time_period", "target_voice", "cur_voice",
+            "download_status", "download_progress",
+            "voice_url", "voice_management_common_params",
+            "map_cruise_position", "flash_light", "pet_snaps",
+            "video_recording_warnings", "visible_to_shared_users",
+            "stream_management_common_params",
+            "stream_management_current_physical_control_lock",
+            "self_check_items", "self_check_results",
+            "ai_object_id", "ai_picture_info",
+            "custom_common_params", "query_id",
+            "updata_difference", "custom_self_check_items",
+            "self_check_result", "spec_test",
+            "low_region_strategy", "precise_diagnosis",
+            "lidar_status", "lidar_button_type", "on",
+        ],
+        "prop_map": {
+            "吸力": "mode",
+            "模式": "mode",
+            "扫拖模式": "sweep_mop_type",
+            "清扫方式": "sweep_type",
+            "拖地水量": "mop_water_output_level",
+            "提示音": "alarm",
+            "音量": "volume",
+            "童锁": "physical_controls_locked",
+            "地毯增压": "carpet_boost",
+            "地毯避让": "carpet_avoidance",
+            "地毯显示": "carpet_display",
+            "断点续扫": "sweep_break_switch",
+            "回洗拖布": "enable_mop_wash",
+            "回洗频率": "frequency_mop_wash",
+            "拖布出水量": "water_output_for_washing_mop",
+            "烘干时长": "drying_time",
+            "自动集尘": "auto_dust_arrest",
+            "集尘频率": "dust_arrest_frequency",
+            "自动烘干拖布": "auto_mop_dry",
+            "自动上下水": "auto_water_change",
+            "使用清洁剂": "use_detergent",
+            "AI清洁": "ai_cleaning",
+            "热水洗拖布": "hot_water_mop_wash",
+            "拖布自动抬升": "mop_auto_lift",
+            "脏污识别": "dirt_detection",
+            "物体识别": "object_detection",
+            "房间识别": "room_detection",
+            "地板材质识别": "floor_material_detection",
+            "勿扰": "no_disturb",
+            "清洁剂自投放": "detergent_self_delivery",
+        },
+        "value_map": {
+            "开": True,
+            "关": False,
+            "开启": True,
+            "关闭": False,
+            "安静": 1,
+            "标准": 2,
+            "强力": 3,
+            "超强": 4,
+            "仅扫地": 1,
+            "仅拖地": 2,
+            "扫拖": 3,
+            "先扫后拖": 4,
+            "全屋清洁": 1,
+            "建图": 2,
+            "选区清扫": 3,
+            "沿边模式": 4,
+            "定制清扫": 5,
+            "指哪扫哪": 6,
+            "自定义清扫": 7,
+            "预约清扫": 8,
+            "联动清扫": 9,
+            "快速建图": 10,
+            "AI托管清扫": 11,
+            "关闭": 0,
+            "1挡": 1,
+            "2挡": 2,
+            "3挡": 3,
+            "按房间": 0,
+            "按面积": 1,
+            "按时间": 2,
+            "深度": 0,
+            "日常": 1,
+            "节水": 2,
+            "2小时": 1,
+            "3小时": 2,
+            "4小时": 3,
+            "一次": 1,
+            "两次": 2,
+            "三次": 3,
+        },
+        "display_map": {
+            "status": "工作状态",
+            "fault": "故障状态",
+            "battery_level": "电池电量",
+            "charging_state": "充电状态",
+            "cleaning_area": "当前清扫面积",
+            "cleaning_time": "当前清扫时长",
+            "clean_times": "清洁次数",
+            "mode": "吸力模式",
+            "sweep_mop_type": "扫拖模式",
+            "sweep_type": "清扫方式",
+            "mop_water_output_level": "拖地水量",
+            "mop_status": "拖布状态",
+            "carpet_boost": "地毯增压模式",
+            "carpet_avoidance": "地毯避让",
+            "carpet_display": "地毯显示",
+            "sweep_break_switch": "断点续扫",
+            "sleep_status": "休眠状态",
+            "location_status": "定位状态",
+            "enable_mop_wash": "回洗拖布",
+            "frequency_mop_wash": "回洗频率",
+            "water_output_for_washing_mop": "拖布出水量",
+            "drying_time": "烘干时长",
+            "auto_dust_arrest": "自动集尘",
+            "dust_arrest_frequency": "集尘频率",
+            "auto_mop_dry": "自动烘干拖布",
+            "auto_water_change": "自动上下水",
+            "use_detergent": "使用清洁剂",
+            "ai_cleaning": "AI感知清洁",
+            "alarm": "提示音",
+            "volume": "音量",
+            "physical_controls_locked": "物理控制锁",
+            "current_physical_control_lock": "物理控制锁状态",
+            "mop_life_level": "拖布剩余寿命",
+            "mop_left_time": "拖布剩余时间",
+            "no_disturb": "勿扰模式",
+            "current_no_disturb": "当前勿扰状态",
+            "brush_life_level": "主刷剩余寿命",
+            "brush_left_time": "主刷剩余时间",
+            "filter_life_level": "滤芯剩余寿命",
+            "filter_left_time": "滤芯剩余时间",
+            "detergent_left_level": "清洁剂剩余量",
+            "detergent_left_time": "清洁剂剩余时间",
+            "dust_bag_life_level": "尘袋剩余寿命",
+            "dust_bag_left_time": "尘袋剩余时间",
+            "drying_progress": "烘干进度",
+            "cleaning_progress": "清扫进度",
+            "dry_left_time": "烘干剩余时长",
+            "statistical_clean_area": "累计清洁面积",
+            "statistical_cleaning_times": "累计清洁次数",
+            "statistical_cleaning_duration": "累计清洁时长",
+        },
+        "detail_writable": [
+            "mode", "sweep_mop_type", "sweep_type",
+            "mop_water_output_level", "carpet_boost",
+            "carpet_avoidance", "carpet_display",
+            "sweep_break_switch", "enable_mop_wash",
+            "frequency_mop_wash", "water_output_for_washing_mop",
+            "drying_time", "auto_dust_arrest", "dust_arrest_frequency",
+            "auto_mop_dry", "auto_water_change", "use_detergent",
+            "ai_cleaning", "alarm", "volume",
+            "physical_controls_locked", "hot_water_mop_wash",
+            "mop_auto_lift", "dirt_detection", "object_detection",
+            "room_detection", "floor_material_detection",
+            "no_disturb", "detergent_self_delivery",
+        ],
+        "detail_readable": [
+            "status", "fault", "battery_level", "charging_state",
+            "cleaning_area", "cleaning_time", "clean_times",
+            "cleaning_progress", "drying_progress", "dry_left_time",
+            "mode", "sweep_mop_type", "sweep_type",
+            "mop_status", "mop_life_level", "mop_left_time",
+            "current_no_disturb",
+            "brush_life_level", "brush_left_time",
+            "filter_life_level", "filter_left_time",
+            "detergent_left_level", "detergent_left_time",
+            "dust_bag_life_level", "dust_bag_left_time",
+        ],
+        "action_map": {
+            "开始清扫": "start_sweep",
+            "暂停": "pause_sweeping",
+            "继续": "continue_sweep",
+            "回充": "start_charge",
+            "开始拖地": "start_mop",
+            "开始扫拖": "start_sweep_mop",
+            "开始建图": "start_build_map",
+            "开始集尘": "start_dust_arrest",
+            "开始洗拖布": "start_mop_wash",
+            "开始烘干": "start_dry",
+            "停止烘干": "stop_dry",
+            "停止洗拖布": "stop_mop_wash",
+            "停止并回充": "stop_and_gocharge",
+        },
+        "detail_actions": [
+            "start_sweep", "pause_sweeping", "continue_sweep",
+            "start_charge", "start_mop", "start_sweep_mop",
+            "start_build_map", "start_dust_arrest", "start_mop_wash",
+            "start_dry", "stop_dry", "stop_mop_wash", "stop_and_gocharge",
+        ],
+        "help_examples": {
+            "吸力": ["安静", "标准", "强力", "超强"],
+            "扫拖模式": ["仅扫地", "仅拖地", "扫拖", "先扫后拖"],
+            "清扫方式": ["全屋清洁", "选区清扫", "沿边模式", "快速建图"],
+            "拖地水量": ["关闭", "1挡", "2挡", "3挡"],
+            "回洗频率": ["按房间", "按面积", "按时间"],
+            "烘干时长": ["2小时", "3小时", "4小时"],
+            "地毯增压": ["开", "关"],
+            "地毯避让": ["开", "关"],
+            "自动集尘": ["开", "关"],
+            "自动烘干拖布": ["开", "关"],
+            "提示音": ["开", "关"],
+            "童锁": ["开", "关"],
+            "勿扰": ["开", "关"],
+        },
+        "action_examples": [
+            "开始清扫", "暂停", "继续", "回充", "开始拖地", "开始扫拖",
+        ],
+        "help_hints": {
+            "音量": "支持 0~100",
+        },
+    },
+
+    # ---- 恒温电水壶 (保留：特有 hidden_props + 加热模式值) ----
+    "xiaomi.kettle.ym3pro": {
+        "category": CATEGORY_WATER_HEATER,
+        "hidden_props": [
+            "knob_sequence", "knob_one", "knob_two", "knob_three",
+            "knob_four", "knob_five", "knob_six",
+            "scene_sequence", "scene_one", "scene_two",
+            "scene_three", "scene_four", "scene_five",
+            "extend_mode_sequence", "extend_mode_one",
+            "extend_mode_two", "extend_mode_three",
+            "extend_mode_four", "extend_mode_five", "extend_mode_six",
+        ],
+        "prop_map": {
+            "开关": "on",
+            "温度": "target_temperature",
+            "加热温度": "target_temperature",
+            "保温温度": "keep_warm_temperature",
+            "自动保温": "auto_keep_warm",
+            "定温时长": "keep_warm_time",
+            "勿扰": "no_disturb",
+            "提壶记忆": "lift_remember_temp",
+            "烧开提醒": "boiling_reminder",
+            "定温提醒": "keep_warm_reminder",
+            "模式": "target_mode",
+            "加热模式": "heat_mode",
+        },
+        "value_map": {
+            "开": True,
+            "关": False,
+            "开启": True,
+            "关闭": False,
+            "待机": 0,
+            "煮沸": 1,
+            "其他": 2,
+            "加热": 3,
+        },
+        "display_map": {
+            "status": "工作状态",
+            "fault": "故障状态",
+            "temperature": "当前水温",
+            "target_temperature": "加热温度",
+            "auto_keep_warm": "自动保温",
+            "keep_warm_temperature": "保温温度",
+            "on": "开关状态",
+            "no_disturb": "勿扰模式",
+            "keep_warm_time": "定温时长",
+            "lift_remember_temp": "提壶记忆定温",
+            "boiling_reminder": "烧开提醒",
+            "keep_warm_reminder": "定温提醒",
+            "kettle_lifting": "提壶状态",
+            "warming_time": "已定温时长",
+            "target_mode": "记忆工作模式",
+            "heat_mode": "加热模式",
+        },
+        "detail_writable": [
+            "on", "target_temperature", "auto_keep_warm",
+            "keep_warm_temperature", "no_disturb", "keep_warm_time",
+            "lift_remember_temp", "boiling_reminder",
+            "keep_warm_reminder", "target_mode", "heat_mode",
+        ],
+        "detail_readable": [
+            "status", "fault", "temperature", "target_temperature",
+            "auto_keep_warm", "keep_warm_temperature",
+            "on", "no_disturb", "keep_warm_time",
+            "kettle_lifting", "warming_time",
+        ],
+        "action_map": {
+            "停止工作": "stop_work",
+            "喝水提醒": "push_water_reminder",
+            "推送喝水提醒": "push_water_reminder",
+        },
+        "detail_actions": ["stop_work", "push_water_reminder"],
+        "help_examples": {
+            "温度": ["70", "85", "95"],
+            "保温温度": ["45", "55", "70"],
+            "定温时长": ["60", "120", "1440"],
+            "自动保温": ["开", "关"],
+            "提壶记忆": ["开", "关"],
+            "烧开提醒": ["开", "关"],
+            "定温提醒": ["开", "关"],
+            "勿扰": ["开", "关"],
+        },
+        "action_examples": ["停止工作", "喝水提醒"],
+        "help_hints": {
+            "温度": "支持 40~99°C",
+            "保温温度": "支持 40~90°C",
+            "定温时长": "支持 60~1440 分钟",
+        },
+    },
+}
